pyproject.toml
[project]
name = "roundtable"
version = "0.1.0"
description = "RAG-powered roundtable discussion with podcast guests"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.4",
    "ollama>=0.1",
    "streamlit>=1.30",
    "llama-index>=0.10",
    "llama-index-embeddings-ollama>=0.1",
    "llama-index-vector-stores-chroma>=0.1",
    "python-dotenv>=1.0",
    "openai>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
